Generate bash commands to process phyloscanner output

This function generates bash commands that combine the various Rscripts in the phyloscanner toolkit

phsc.cmd.process.phyloscanner.output.in.directory(tmp.dir, file.patients,
  pty.args)

Arguments

tmp.dir
Directory with phyloscanner output.

file.patients
File name of the file that contains the list of unique individuals/units that the bam files correspond to, possibly after re-naming.

pty.args
List of phyloscanner input variables.

Value

character string of bash commands.