Find pairs of individuals between whom linkage is not excluded phylogenetically

This function identifies pairs of individuals between whom linkage is not excluded phylogenetically in a large number of phyloscanner analyses, and provides detailed information on them. Three R objects are generated: 'linked.pairs' is a data.table that describes pairs of individuals between whom linkage is not excluded phylogenetically. 'relationship.counts' is a data.table that summarises the phylogenetic relationship counts for each pair. 'windows' is a data.table that describes the basic phyloscanner statistics (distance, adjacency, paths between subgraphs) in each deep-sequence phylogeny for each pair.

phsc.find.linked.pairs(indir, batch.regex = "^ptyr([0-9]+)_.*",
  conf.cut = 0.6, neff.cut = 3, verbose = TRUE, dmeta = NULL)

Arguments

indir
Full directory path to output of phyloscanner runs

batch.regex
Regular expression that identifies the batch ID of multiple phyloscanner analyses. Default: '^ptyr([0-9]+)_.*'.

conf.cut
Threshold on the proportion of deep-sequence phylogenies with distant/disconnected subgraphs above which pairs are considered phylogenetically unlinked. Default: 0.6

neff.cut
Threshold on the minimum number of deep-sequence phylogenies with sufficient reads from two individuals to make any phylogenetic inferences. Default: 3.

verbose
Flag to switch on/off verbose mode. Default: TRUE.

dmeta
Optional individual-level meta-data that is to be added to output. Can be NULL.

Value

list of three R objects 'linked.pairs', 'relationship.counts', 'windows'. See description.