Get transmission assignments per window

This function extracts all transmission assignments per window for two individuals.

phsc.get.assignments.by.window.for.couple(id1, id2, infiles)

Arguments

id1
Identifier of the first individual. This is the entry for that individual in the phyloscanner .bam file.

id2
Identifier of the second individual.

infiles
One or more file names of transmission stats rda files. These end typically with '_trmStatsPerWindow.rda'.

Value

data.table with columns ID1 ID2 W_FROM (window start) W_TO (window end), and further columns for each infile that give the transmission assignments per window.