Plot short read phylogenies and highlight individuals

This function plots short read phylogenies and highlights the clades of two individuals in red and blue.

phsc.plot.phy.selected.individuals(phs, dfs, ids,
  plot.cols = rainbow(length(ids)), plot.file = NA, group.redo = FALSE,
  drop.less.than.n.ids = 2, drop.blacklisted = FALSE, pdf.h = 50,
  pdf.rw = 10, pdf.ntrees = 20, pdf.title.size = 40)

Arguments

phs
List of trees in ape format

dfs
data.table with mandatory column 'IDX' and optional column 'TITLE'. IDX is the index of all phylogenies in 'phs' that are to be plotted. TITLE is a title for each sub-plot, for example specifying a window.

ids
Vector of regular expressions that identify individuals to be highlighted in colour.

plot.cols
Vector of colours for each individual

plot.file
If not missing, the phylogenies will be printed to file.

group.redo
Logical, indicating if the colour groups should be recalculated from ids.

drop.blacklisted
Logical, indicating if all blacklisted taxa should be dropped prior to plotting.

pdf.h
Height of the pdf file in inches.

pdf.rw
Relative width of the pdf file, internally multiplied by the number of phylogenies to give the total width in inches.

pdf.ntrees
Number of trees per pdf.

pdf.title.size
Size of pdf title in inches.

Value

List of ggtree objects, ready for printing.