Read processed phyloscanner output

This function creates R data.tables from processed phyloscanner output, for further data analysis in R.

phsc.read.processed.phyloscanner.output.in.directory(prefix.infiles,
  save.file.base, read.likelytransmissions = TRUE, read.trees = TRUE,
  read.subtrees = TRUE, resume = FALSE, zip = FALSE)

Arguments

prefix.infiles
Full path name to processed phyloscanner output

save.file.base
Output will be stored to files that start with 'save.file.base'.

read.likelytransmissions
If TRUE, read and process likely transmissions

read.trees
If TRUE, read and process trees

read.subtrees
If TRUE, read and process subtree files

resume
If TRUE, the function does not process existing rda files.

zip
If TRUE, the function zips processed phyloscanner output, and then deletes the zipped, processed phyloscanner output files.

Value

Nothing, rda objects written to file.